refactor: move RuleConfig class
