+++ src/main/java/co/fineants/api/domain/validator/domain/Validator.java
@@ -1,4 +1,4 @@
-package co.fineants.api.domain.validator;
+package co.fineants.api.domain.validator.domain;
 
 public interface Validator<T> {
 	void validate(T target);
