+++ src/main/java/co/fineants/api/domain/validator/domain/member/SignUpValidator.java
@@ -1,11 +1,11 @@
-package co.fineants.api.domain.validator.member;
+package co.fineants.api.domain.validator.domain.member;
 
 import java.util.Arrays;
 import java.util.List;
 
 import co.fineants.api.domain.member.domain.entity.Member;
-import co.fineants.api.domain.validator.MemberValidationRule;
-import co.fineants.api.domain.validator.Validator;
+import co.fineants.api.domain.validator.domain.MemberValidationRule;
+import co.fineants.api.domain.validator.domain.Validator;
 
 public class SignUpValidator implements Validator<Member> {
 	private final List<MemberValidationRule> rules;
