+++ src/test/java/co/fineants/api/domain/member/domain/rule/EmailValidatorTest.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import co.fineants.AbstractContainerBaseTest;
 import co.fineants.api.domain.member.domain.entity.Member;
 import co.fineants.api.domain.member.repository.MemberRepository;
-import co.fineants.api.domain.validator.member.EmailValidator;
+import co.fineants.api.domain.validator.domain.member.EmailValidator;
 import co.fineants.api.global.errors.exception.business.EmailDuplicateException;
 
 class EmailValidatorTest extends AbstractContainerBaseTest {
