+++ src/test/java/co/fineants/api/domain/member/domain/rule/NicknameValidatorTest.java
@@ -9,7 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import co.fineants.AbstractContainerBaseTest;
 import co.fineants.api.domain.member.repository.MemberRepository;
-import co.fineants.api.domain.validator.member.NicknameValidator;
+import co.fineants.api.domain.validator.domain.member.NicknameValidator;
 import co.fineants.api.global.errors.exception.business.NicknameDuplicateException;
 import co.fineants.api.global.errors.exception.business.NicknameInvalidInputException;
